1. Fixed PSQLADM-293
Cause: When all readers are removed from original HG the writer is
promoted to the reader, however pushNewNode() tried to find the pushed
node in readers (which was failing). That caused creatition of action
with empty node, (DNS) which caused out of bound panic.

Fix: removed unnecessary node getting from BackupReaders. The goal is to
remove the node from maintenance readers first, then add it as the
reader.
Tests introduced:
reader_add1.test
reader_add2.test

2. Removed dead code (command line args implementation attempt)

3. Fixed fmt.Printf() formating

4. MTR tests use common configurations (so far 2 config files which
differ with WriterIsAlsoReader configuration)