Add support for ProxySQL cluster  we will check if the node were we are has a lock or if can acquire one.  If not we will return nil to indicate program must exit given either there is already another node holding the lock  or this node is not in a good state to acquire a lock  Outside call To get and check the active list of ProxySQL server we call ProxySQLCLuster.GetProxySQLnodes  All the DB operations are done connecting locally to the ProxySQL node running the scheduler
Find lock method review all the nodes existing in the Proxysql for an active LOck
it checks only nodes that are reachable.
Checks for:
	- existing log locally
	- lock on another node
	- lock time comparing it with lockclustertimeout parameter

Interestingly ProxySQL has not clue if a ProxySQL nodes ur down. Or at least is not reported in the proxysql_server tables or any stats table
Given that we check if nodes are reachable opening a connection and closing it